Refactor code and rename some files/folders
