+++ minishell/src/check_input/env_variables.c
