+++ minishell/src/check_input/redirections_space.c
