+++ minishell/src/main.c
@@ -6,7 +6,7 @@
 /*   By: bszabo <bszabo@student.42vienna.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/09 13:42:31 by bszabo            #+#    #+#             */
-/*   Updated: 2024/04/10 16:23:17 by bszabo           ###   ########.fr       */
+/*   Updated: 2024/04/13 09:58:15 by bszabo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,9 +29,7 @@ static int	main_loop(t_data *data)
 			add_history(data->line);
 			if (check_line(data) == ERROR)
 				continue;
-			if (parse_line(data) == ERROR)
-				continue;
-			if (execute(data) == ERROR)
+			if (parse_execute_line(data) == ERROR)
 				continue;
 		}
 	}
