+++ minishell/src/parse_execute/cmd_data.c
@@ -6,7 +6,7 @@
 /*   By: bszabo <bszabo@student.42vienna.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/29 19:40:17 by bszabo            #+#    #+#             */
-/*   Updated: 2024/04/12 08:45:06 by bszabo           ###   ########.fr       */
+/*   Updated: 2024/04/13 10:00:19 by bszabo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
