+++ minishell/src/parse_execute/parse_execute.c
@@ -6,7 +6,7 @@
 /*   By: bszabo <bszabo@student.42vienna.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/25 16:27:38 by bszabo            #+#    #+#             */
-/*   Updated: 2024/04/06 06:37:29 by bszabo           ###   ########.fr       */
+/*   Updated: 2024/04/13 10:02:41 by bszabo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ static int	split_and_count(t_data *data)
 
 // parse the line, prepare data for execution
 // return ERROR if any of the steps fail, OK otherwise
-int	parse_line(t_data *data)
+int	parse_execute_line(t_data *data)
 {
 	if (split_and_count(data) == ERROR)
 		return (ERROR);
